+++ src/contracts/index.ts
@@ -62,3 +62,4 @@ export * from "./EventContractSource";
 export * from "./EventContractCollectionRunner";
 export * from "./EventContractCollectionRunnerRecoveryControl";
 export * from "./EventContractCollectionRunnerRuntime";
+export * from "./EventContractCollectionRunnerRuntimeAssembly";