+++ src/engines/event-contract-collection-runner-runtime/index.ts
@@ -2,3 +2,4 @@ export * from "./EventContractCollectionRunnerRuntimeFoundation";
 export * from "./EventContractCollectionRunnerScheduler";
 export * from "./EventContractCollectionRunnerFixtureWorker";
 export * from "./EventContractCollectionRunnerRuntimeOperator";
+export * from "./EventContractCollectionRunnerRuntimeAssemblyPlanner";